requirements.txt
py-clob-client
python-dotenv
websockets
requests
